+++ _infra/helm/collection-instrument/templates/deployment.yaml
@@ -261,8 +261,6 @@ spec:
             {{- end }}
           - name: RABBITMQ_AMQP_SURVEY_RESPONSE
             value: "amqp://$(SDX_RABBITMQ_USERNAME):$(SDX_RABBITMQ_PASSWORD)@$(SDX_RABBITMQ_HOST):$(SDX_RABBITMQ_PORT)"
-          - name: SAVE_SEFT_IN_GCP
-            value: "{{ .Values.saveSeftInGcp }}"
           - name: ONS_GNU_FINGERPRINT
             value: "{{ .Values.gnu.fingerprint }}"
           - name: ONS_GNU_RECIPIENT
