+++ _infra/helm/collection-instrument/values.yaml
@@ -68,8 +68,6 @@ gcp:
     prefix: ""
   topic: topic
 
-saveSeftInGcp: false
-
 gnu:
   recipient: dev@example.com
   fingerprint: ED1B7A3EADF95687
