+++ application/controllers/survey_response.py
@@ -10,10 +10,6 @@ from application.controllers.helper import (
     is_valid_file_extension,
     is_valid_file_name_length,
 )
-from application.controllers.rabbit_helper import (
-    initialise_rabbitmq_queue,
-    send_message_to_rabbitmq_queue,
-)
 from application.controllers.service_helper import (
     get_business_party,
     get_case_group,
@@ -70,11 +66,6 @@ class SurveyResponse(object):
                 raise SurveyResponseError()
 
     @staticmethod
-    def initialise_messaging():
-        log.info("Initialising rabbitmq queue for Survey Responses", queue=RABBIT_QUEUE_NAME)
-        return initialise_rabbitmq_queue(RABBIT_QUEUE_NAME)
-
-    @staticmethod
     def _create_json_message_for_file(file_name, file, case_id, survey_ref):
         """
         Create json message from file
@@ -119,61 +110,7 @@ class SurveyResponse(object):
 
         return True, ""
 
-    def get_file_name_and_survey_ref(self, case_id, file_extension):
-        """
-        Generate the file name for the upload, if an external service can't find the relevant information
-        a None is returned instead.
 
-        .. note:: returns two seemingly disparate values because the survey_ref is needed for filename anyway,
-            and resolving requires calls to http services, doing it in one function minimises network traffic.
-            survey_id as returned by collection exercise is a uuid, this is resolved by a call to
-            survey which returns it as surveyRef which is the 3 digit id that other services refer to as survey_id
-
-        :param case_id: The case id of the upload
-        :param file_extension: The upload file extension
-        :return: file name and survey_ref or None
-        """
-
-        log.info("Generating file name", case_id=case_id)
-
-        case_group = get_case_group(case_id)
-        if not case_group:
-            return None, None
-
-        collection_exercise_id = case_group.get("collectionExerciseId")
-        collection_exercise = get_collection_exercise(collection_exercise_id)
-        if not collection_exercise:
-            return None, None
-
-        exercise_ref = collection_exercise.get("exerciseRef")
-        survey_id = collection_exercise.get("surveyId")
-        survey_ref = get_survey_ref(survey_id)
-        if not survey_ref:
-            return None, None
-
-        ru = case_group.get("sampleUnitRef")
-        exercise_ref = self._format_exercise_ref(exercise_ref)
-
-        business_party = get_business_party(
-            case_group["partyId"], collection_exercise_id=collection_exercise_id, verbose=True
-        )
-        if not business_party:
-            return None, None
-        check_letter = business_party["checkletter"]
-
-        time_date_stamp = time.strftime("%Y%m%d%H%M%S")
-        file_name = "{ru}{check_letter}_{exercise_ref}_{survey_ref}_{time_date_stamp}{file_format}".format(
-            ru=ru,
-            check_letter=check_letter,
-            exercise_ref=exercise_ref,
-            survey_ref=survey_ref,
-            time_date_stamp=time_date_stamp,
-            file_format=file_extension,
-        )
-
-        log.info("Generated file name for upload", filename=file_name)
-
-        return file_name, survey_ref
 
     @staticmethod
     def check_if_file_size_too_small(file_size):
